# Typed Quad values