--- a/graph/bolt/all_iterator.go
+++ b/graph/bolt/all_iterator.go
@@ -42,7 +42,7 @@ type AllIterator struct {
 
 func NewAllIterator(bucket []byte, d quad.Direction, qs *QuadStore) *AllIterator {
 	return &AllIterator{
-		nodes:  d == 0,
+		nodes:  d == quad.Any,
 		uid:    iterator.NextUID(),
 		bucket: bucket,
 		dir:    d,