--- a/graph/bolt/iterator.go
+++ b/graph/bolt/iterator.go
@@ -224,45 +224,45 @@ func PositionOf(tok *Token, d quad.Direction, qs *QuadStore) int {
 		case quad.Subject:
 			return 0
 		case quad.Predicate:
-			return hashSize
+			return quad.HashSize
 		case quad.Object:
-			return 2 * hashSize
+			return 2 * quad.HashSize
 		case quad.Label:
-			return 3 * hashSize
+			return 3 * quad.HashSize
 		}
 	}
 	if bytes.Equal(tok.bucket, posBucket) {
 		switch d {
 		case quad.Subject:
-			return 2 * hashSize
+			return 2 * quad.HashSize
 		case quad.Predicate:
 			return 0
 		case quad.Object:
-			return hashSize
+			return quad.HashSize
 		case quad.Label:
-			return 3 * hashSize
+			return 3 * quad.HashSize
 		}
 	}
 	if bytes.Equal(tok.bucket, ospBucket) {
 		switch d {
 		case quad.Subject:
-			return hashSize
+			return quad.HashSize
 		case quad.Predicate:
-			return 2 * hashSize
+			return 2 * quad.HashSize
 		case quad.Object:
 			return 0
 		case quad.Label:
-			return 3 * hashSize
+			return 3 * quad.HashSize
 		}
 	}
 	if bytes.Equal(tok.bucket, cpsBucket) {
 		switch d {
 		case quad.Subject:
-			return 2 * hashSize
+			return 2 * quad.HashSize
 		case quad.Predicate:
-			return hashSize
+			return quad.HashSize
 		case quad.Object:
-			return 3 * hashSize
+			return 3 * quad.HashSize
 		case quad.Label:
 			return 0
 		}
@@ -298,7 +298,7 @@ func (it *Iterator) Size() (int64, bool) {
 func (it *Iterator) Describe() graph.Description {
 	return graph.Description{
 		UID:       it.UID(),
-		Name:      it.qs.NameOf(&Token{it.bucket, it.checkID}),
+		Name:      it.qs.NameOf(&Token{it.bucket, it.checkID}).String(),
 		Type:      it.Type(),
 		Tags:      it.tags.Tags(),
 		Size:      it.size,