### Strictly-typed quad values

The idea is to change string type of quad directions to interface-based Value, to allow to preserve types of subject and object particularly.

```Go
type Value interface {
	String() string
}

type Quad struct {
	Subject   Value
	Predicate Value
	Object    Value
	Label     Value
}
```

As a part of effort, unify tests for QuadStore to check that all implementations behaves in a similar way.

### Supported types:
```Go
// Raw is a Turtle/NQuads-encoded value.
type Raw string

// String is an RDF string value (ex: "name").
type String string

// TypedString is an RDF value with type (ex: "name"^^<type>).
type TypedString struct {
	Value String
	Type  IRI
}

// LangString is an RDF string with language (ex: "name"@lang).
type LangString struct {
	Value String
	Lang  string
}

// IRI is an RDF Internationalized Resource Identifier (ex: <name>).
type IRI string

// BNode is an RDF Blank Node (ex: _:name).
type BNode string

// Int is a native wrapper for int64 type.
type Int int64

// Float is a native wrapper for float64 type.
type Float float64

// Bool is a native wrapper for bool type.
type Bool bool

// Time is a native wrapper for time.Time type.
type Time time.Time
```
### TODO:
- [x] Value definition
- [x] Update cquads and nquads parsers and tests
- [x] Update QS and tests
- [x] Update integration tests
- [x] Unified tests for QS
- [x] Memstore support
- [x] Bolt support
- [x] LevelDB support (switch to protobuf)
- [x] Mongo support (untyped)
- [x] SQL support (untyped)
- [x] GAE support
- [x] Support for native types (int, float, etc)
- [x] Typed values in Mongo
- [x] Typed values in SQL
- [ ] Typed values in GAE
- [x] Path lib support
- [x] Gremlin support
- [x] Recognize TypedString and LandString in cquads parser
- [x] Registration for TypedString converters
- [ ] Support for custom types
- [ ] Migration code for SQL

#### Problems so far:
- [ ] goprotobuf will not work on GAE, thus we have no support for typed values for now

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cayleygraph/cayley/378)
<!-- Reviewable:end -->