@dennwc 2016-03-06
Performance improvements for leveldb (d47e03d42637be8ef9c12685dcee4561958e6706):
```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkNamePredicate-8                      908541        764269        -15.88%
BenchmarkLargeSetsNoIntersection-8            152631283     102180698     -33.05%
BenchmarkVeryLargeSetsSmallIntersection-8     370620517     249059029     -32.80%
BenchmarkHelplessContainsChecker-8            677294497     460773274     -31.97%
BenchmarkHelplessNotContainsFilms-8           660483822     436792429     -33.87%
BenchmarkHelplessNotContainsActors-8          688224013     467125431     -32.13%
BenchmarkNetAndSpeed-8                        21529498      15497605      -28.02%
BenchmarkKeanuAndNet-8                        10909581      8805261       -19.29%
BenchmarkKeanuAndSpeed-8                      14575400      11968996      -17.88%
BenchmarkKeanuOther-8                         89169973      62986376      -29.36%
BenchmarkKeanuBullockOther-8                  147350741     130021983     -11.76%
BenchmarkSaveBogartPerformances-8             4887044       3698388       -24.32%
```