@dennwc 2016-03-07
### TODO list:
- [ ] GAE support
- [ ] Typed values in Mongo
- [ ] Typed values in SQL
- [ ] Path lib support
- [ ] Gremlin support
- [ ] MQL/SExp support?
- [ ] Recognize TypedString and LandString in cquads parser
- [ ] Registration for TypedString converters