@dennwc 2016-03-20
Commit 6ae2eaccec0fd79fda87a2f5ce5e44f3bece3437 fixes #290. The syntax:
```js
g.V().Filter( gt(10) ).All()
```