@kortschak 2016-07-09
Nice plan.