@barakmich 2016-07-15
Reviewed 10 of 56 files at r1, 29 of 78 files at r2, 3 of 14 files at r3.
Review status: 40 of 95 files reviewed at latest revision, 4 unresolved discussions.

---

*[graph/bolt/all_iterator.go, line 45 \[r3\]](https://reviewable.io:443/reviews/cayleygraph/cayley/378#-KMgbOAoVxzoNh7Tay4N:-KMgbOApl5vjeOW1GwW_:-1509546067) ([raw file](https://github.com/cayleygraph/cayley/blob/b99aa1dab33f82f6f3a16d1f8f81e77492f73fd8/graph/bolt/all_iterator.go#L45)):*
> ```Go
> func NewAllIterator(bucket []byte, d quad.Direction, qs *QuadStore) *AllIterator {
> 	return &AllIterator{
> 		nodes:  d == 0,
> ```

`d == quad.Any`?

---

*[graph/path/path.go, line 75 \[r3\]](https://reviewable.io:443/reviews/cayleygraph/cayley/378#-KMgelIsQnJftWttqjBs:-KMgelIsQnJftWttqjBt:-1340118616) ([raw file](https://github.com/cayleygraph/cayley/blob/b99aa1dab33f82f6f3a16d1f8f81e77492f73fd8/graph/path/path.go#L75)):*
> ```Go
> 
> // StartMorphism creates a new Path with no underlying QuadStore.
> func StartMorphism(nodes ...quad.Value) *Path {
> ```

I'm of multiple minds about this. It breaks current API users. On the other hand, they're likely vendoring us. 

I'm open to suggestions, and not strictly against it as written, but we should at least discuss.  I'd lean toward making these `...interface{}` and type assert between strings and quad.Value for these cases, exactly as done in `morphism_apply_functions.go`

(Same also for the other ones in this change)

---

*[graph/proto/serializations.proto, line 66 \[r3\]](https://reviewable.io:443/reviews/cayleygraph/cayley/378#-KMggn94CsiCma5CeMVJ:-KMggn95tMtjo_Fb1s2r:726469523) ([raw file](https://github.com/cayleygraph/cayley/blob/b99aa1dab33f82f6f3a16d1f8f81e77492f73fd8/graph/proto/serializations.proto#L66)):*
> ```protobuf
>     int32 nanos = 2;
>   }
>   oneof value {
> ```

:clap: 

---

*[quad/value.go, line 62 \[r3\]](https://reviewable.io:443/reviews/cayleygraph/cayley/378#-KMjBzopEbUZHmK9lxDP:-KMjBzoqrwS-oedRHJSq:-10665999) ([raw file](https://github.com/cayleygraph/cayley/blob/b99aa1dab33f82f6f3a16d1f8f81e77492f73fd8/quad/value.go#L62)):*
> ```Go
> type Raw string
> 
> func (s Raw) String() string { return string(s) }
> ```

After tracing down why the query language had to change --- `g.V('"Humphrey Bogart"')`, notice extra quotes, I came back here.

String() is being conflated with, effectively, MarshalRDF().

Sometimes these are the same. But in query languages, we'd rather interact with their string-ish representations, not their RDF-ish representations -- unless they *are* the same.

TL;DR, an RDF() method, used mainly in this package, and the String() function should work okay in the others.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cayleygraph/cayley/378)*
<!-- Sent from Reviewable.io -->