@dennwc 2016-07-15
Review status: 40 of 95 files reviewed at latest revision, 3 unresolved discussions.

---

*[graph/path/path.go, line 75 \[r3\]](https://reviewable.io:443/reviews/cayleygraph/cayley/378#-KMgelIsQnJftWttqjBs:-KMjRcHWp6j6Z0V9pE5G:-598674595) ([raw file](https://github.com/cayleygraph/cayley/blob/b99aa1dab33f82f6f3a16d1f8f81e77492f73fd8/graph/path/path.go#L75)):*
<details><summary><i>Previously, barakmich (Barak Michener) wrote…</i></summary>
> I'm of multiple minds about this. It breaks current API users. On the other hand, they're likely vendoring us. 
> 
> I'm open to suggestions, and not strictly against it as written, but we should at least discuss.  I'd lean toward making these `...interface{}` and type assert between strings and quad.Value for these cases, exactly as done in `morphism_apply_functions.go`
> 
> (Same also for the other ones in this change)
</details>
This was a hard decision, in fact. `interface{}` is nice because you can seamlessly pass `string`, `int` and other values. But it also hides any potential errors. Also, each function must call some helper that casts values properly. I decided it's better to let user cast it once on his side and make our code cleaner. But I'm still in doubt.

---

*[quad/value.go, line 62 \[r3\]](https://reviewable.io:443/reviews/cayleygraph/cayley/378#-KMjBzopEbUZHmK9lxDP:-KMjK-zuNcgHDQxX_MQ7:131596061) ([raw file](https://github.com/cayleygraph/cayley/blob/b99aa1dab33f82f6f3a16d1f8f81e77492f73fd8/quad/value.go#L62)):*
<details><summary><i>Previously, barakmich (Barak Michener) wrote…</i></summary>
> After tracing down why the query language had to change --- `g.V('"Humphrey Bogart"')`, notice extra quotes, I came back here.
> 
> String() is being conflated with, effectively, MarshalRDF().
> 
> Sometimes these are the same. But in query languages, we'd rather interact with their string-ish representations, not their RDF-ish representations -- unless they *are* the same.
> 
> TL;DR, an RDF() method, used mainly in this package, and the String() function should work okay in the others.
</details>
Yes, I also like an idea to separate RDF marshaling/unmarshaling into it's own method set.

But the question regarding `g.V('"Humphrey Bogart"')` is more specific.  It works this way: any string is assumed to be in raw RDF (quad.Raw), thus it calculates hash of RDF form and then saves it in iterator. QuadStore will use hash to do lookup and it may hit quad.String node `Humphrey Bogart` (without quotes), because its RDF representation is the same as `"Humphrey Bogart"` (quad.Raw value). In this case queries like `g.V("<bob>")` and `g.V('"Bob"')` works, and they also works with helpers: `g.V(iri("bob"))`, `g.V(str("Bob"))`. I agree that this is too verbose. The alternative can be following: assume that strings are of quad.String type, and make helper for RDF form. In this case only `g.V(iri("bob"))` and `g.V(raw("<bob>"))` will work, but for strings it would be nicer: `g.V("Bob")` and `g.V(raw('"Bob"'))`.

And it'a also leads to broader discussion. Are nodes `10`, `"10"` and `"10"^^<int>` the same? What form should be hashed? May it use these forms interchangeably? 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cayleygraph/cayley/378)*
<!-- Sent from Reviewable.io -->